refactor: EventPeriodRequest -> CreateEventPeriodRequest 로 변경
